try to add the dtype option

+++ tests/python/unittest/test_ndarray.py
@@ -727,7 +727,7 @@ def test_order():
 
     # test topk with a big shape
     a = mx.nd.arange(0, 54686454, step=1, repeat=1, dtype=np.int32)
-    assert_almost_equal(a.topk(k=54686454).asnumpy(), a.asnumpy()[::-1])
+    assert_almost_equal(a.topk(k=54686454, dtype=np.int32).asnumpy(), a.asnumpy()[::-1])
     a.attach_grad()
 
     k = 10
